requirements.txt
tornado==6.4.2
selenium==3.141.0
requests==2.23.0
lxml==5.3.1
redis==3.4.1
sqlalchemy==1.3.18
pymysql==0.10.0
APScheduler==3.6.3
gevent==24.11.1
beautifulsoup4==4.4.0
pandas==2.2.3
jieba==0.42.1
pyjwt==2.0.1
python3-pika==0.9.14
celery==5.2.6
pydantic==2.10.5
httpx==0.28.1
openai==1.70.0
cryptography==43.0.3
pyyaml==6.0.2
typing==3.7.4.3
uuid==1.30
openpyxl==3.1.5

